+++ test/runtests.jl
@@ -4,10 +4,15 @@ using Test
 @testset "LineCableModels.jl tests" begin
 	@info "Running BaseParams tests..."
 	@testset "BaseParams module" begin
-		# This line executes all the tests defined in baseparams.jl
 		include("baseparams.jl")
 	end
 	@info "BaseParams tests completed."
+
+	@testset "DataModel module 1/1" begin
+		include("datamodel.jl")
+	end
+	@info "DataModel tests completed."
+
 
 	@info "All tests completed."
 end